reuse cost matrix builder code
