+++ lindera-dictionary-builder/src/compress.rs
@@ -1,0 +1,34 @@
+use std::io::Write;
+
+#[cfg(feature = "compress")]
+use lindera_compress::compress;
+use lindera_core::error::LinderaErrorKind;
+use lindera_core::LinderaResult;
+use lindera_decompress::Algorithm;
+
+#[cfg(feature = "compress")]
+pub fn compress_write<W: Write>(
+    buffer: &[u8],
+    algorithm: Algorithm,
+    writer: &mut W,
+) -> LinderaResult<()> {
+    let compressed = compress(buffer, algorithm)
+        .map_err(|err| LinderaErrorKind::Compress.with_error(anyhow::anyhow!(err)))?;
+    bincode::serialize_into(writer, &compressed)
+        .map_err(|err| LinderaErrorKind::Io.with_error(anyhow::anyhow!(err)))?;
+
+    Ok(())
+}
+
+#[cfg(not(feature = "compress"))]
+pub fn compress_write<W: Write>(
+    buffer: &[u8],
+    _algorithm: Algorithm,
+    writer: &mut W,
+) -> LinderaResult<()> {
+    writer
+        .write_all(buffer)
+        .map_err(|err| LinderaErrorKind::Io.with_error(anyhow::anyhow!(err)))?;
+
+    Ok(())
+}
