+++ lindera-dictionary-builder/src/lib.rs
@@ -1,3 +1,6 @@
-mod user_dict;
+pub mod compress;
+pub mod cost_matrix;
+pub mod user_dict;
 
+pub use cost_matrix::{CostMatrixBuilder, CostMatrixBuilderOptions, CostMatrixBuilderOptionsError};
 pub use user_dict::{UserDictBuilder, UserDictBuilderOptions, UserDictBuilderOptionsError};
